% Generated by roxygen2 (4.0.1): do not edit by hand
\name{auxfun}
\alias{auxfun}
\alias{fullweek}
\alias{mkpar}
\alias{monday}
\alias{mondayf}
\alias{mondays}
\alias{na.0}
\alias{namax}
\alias{namean}
\alias{namin}
\alias{numdate}
\alias{seasday}
\alias{to.cels}
\alias{to.na.0}
\title{Auxilliary functions for tempmomo}
\usage{
na.0(a, b = 0)

to.na.0(a, b = 0)

to.cels(a)

numdate(a)

seasday(a)

namax(a)

namin(a)

namean(a)

monday(a, ...)

mondays(first, last, ...)

mondayf(a, first, last, ...)

fullweek(a, type = c("first", "last"), ...)

mkpar(n, row = TRUE)
}
\description{
Various little helper functions, not exported, for use within the package

min and max with NA:s removed

create a list of full weeks between dates

create a factor variable of full series of mondays

give the monday of first (last) full week

create pertty grid of n frames
}

