% Generated by roxygen2 (4.0.1): do not edit by hand
\name{euroMoMoTemp}
\alias{euroMoMoTemp}
\title{A package to download weather data from various sources}
\description{
Main use is to provide weather data to the EuroMoMo project \URL{www.euromomo.eu}
}
\details{
Documentation for importing data from the NOAA site
\url{ftp://ftp.ncdc.noaa.gov/pub/data/gsod/}
Data description: \url{ftp://ftp.ncdc.noaa.gov/pub/data/gsod/GSOD_DESC.txt}
Lists of sites  : \url{ftp://ftp.ncdc.noaa.gov/pub/data/gsod/ish-history.csv}
}

