% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getEuroMap}
\alias{getEuroMap}
\title{Download map data from EuroStat}
\usage{
getEuroMap(year = "2010", resolution = "03M", force = FALSE,
  basepath = NULL)
}
\arguments{
\item{year}{The year of NUTS regions (Use with caution)}

\item{resolution}{The resolution of the shapefile (use with caution)}

\item{force}{Should the file be downloaded even if it already exists}

\item{basepath}{The directory where the files should be downloaded}
}
\value{
None
}
\description{
Downloads NUTS regions from EuroStat website at \url{http://epp.eurostat.ec.europa.eu/cache/GISCO/geodatafiles/NUTS_2010_03M.zip}
}

