% Generated by roxygen2 (4.0.1): do not edit by hand
\name{noaaGetCountry}
\alias{noaaGetCountry}
\title{Get all data from all stations in a country}
\usage{
noaaGetCountry(usecountry = "FI", years = 2008:2013, sitesdata = sites,
  force = FALSE, thisy = TRUE, basepath = NULL, uplag = 0,
  countrycode = c("country", "NUTS", "NUTS0i", "NUTS3i"))
}
\arguments{
\item{usecountry}{the name of the country as listed in ftp://ftp.ncdc.noaa.gov/pub/data/gsod/country-list.txt}

\item{years}{vector of years to use}

\item{sitesdata}{a data frame created using noaaGetSites}

\item{force}{Should the data be downloaded even if it exists in the cache}

\item{thisy}{Should this year's potentially incomplete data be downloaded even if it exists in the cache}

\item{basepath}{Directory where the data is downloaded (in subdirectory download)}

\item{uplag}{How old files should be updated}

\item{countrycode}{What variable in the sites -file gives the country name}
}
\value{
A data frame with all stations and all years data
}
\description{
Go through all stations in the country by each year and download the data
}

