% Generated by roxygen2 (4.0.1): do not edit by hand
\name{noaaGetSites}
\alias{noaaGetSites}
\title{Download the station specific metadata from the NOAA site}
\usage{
noaaGetSites(force = FALSE)
}
\arguments{
\item{force}{should the file be downloaded or use the cached file?}
}
\description{
Downloads the list of stations, reads it in and cleans the variables
}

