% Generated by roxygen2 (4.0.1): do not edit by hand
\name{noaaGetSiteYear}
\alias{noaaGetSiteYear}
\title{Downloads data for one station and year}
\usage{
noaaGetSiteYear(site, year = 2013, force = FALSE, thisy = TRUE,
  uplag = 0, basepath = NULL)
}
\arguments{
\item{site}{the identificator of the station, must be six letter string (i.e. with trailing 0's if shorter)}

\item{year}{the year whose data is to be downloaded}

\item{force}{should the data be downloaded even if it is found in the cache (default=FALSE)}

\item{thisy}{should this years data be downloaded even if force=FALSE}

\item{uplag}{how old files should be updated, days}

\item{basepath}{where the cached files should go}
}
\value{
a dataframe with single years data from a single station
}
\description{
The NOAA site has each station's data separated into files. These files are further split into years with one directory for each year.
The data is downloaded, read in and transformed to use metric measurement units
}

