% Generated by roxygen2 (4.0.1): do not edit by hand
\name{noaaMapRegions}
\alias{noaaMapRegions}
\title{Match sites to map areas}
\usage{
noaaMapRegions(sites, map, level = 0, type = c("inarea", "distance",
  "bestquess"))
}
\arguments{
\item{sites}{data frame with lat&lon}

\item{map}{ShapeSpatialDataFrame with the regions}

\item{level}{Level of area}

\item{type}{What kind of match should be returned}
}
\value{
A matched label of the region
}
\description{
Match sites to map areas
}

